Simple Linear Regression
Problem  Statement:                                                                                                                     
The Company has the data of employees with years of experience and salary.                                                     
The goal is to predict the new employee with the number of years exp should get what amount of Salary should he get.                
Understanding the correlation between the Years of Experience and Salary(Independent and Dependent).